requirements.txt
annotated-types==0.6.0
anyio==4.1.0
bardapi==0.1.38
beautifulsoup4==4.12.2
blinker==1.7.0
browser-cookie3==0.19.1
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
deep-translator==1.11.4
deprecation==2.1.0
exceptiongroup==1.2.0
firebase-admin==6.3.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
google-api-core==2.14.0
google-api-python-client==2.111.0
google-auth==2.23.4
google-auth-httplib2==0.2.0
google-cloud-core==2.3.3
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-cloud-translate==3.12.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.61.0
gotrue==1.3.0
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
hyperframe==6.0.1
idna==3.6
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
langdetect==1.0.9
lz4==4.3.2
MarkupSafe==2.1.3
msgpack==1.0.7
packaging==23.2
Pillow==10.1.0
postgrest==0.13.0
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
realtime==1.0.0
replicate==0.21.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
storage3==0.7.0
StrEnum==0.4.15
supabase==2.1.0
supafunc==0.3.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
websockets==10.4
Werkzeug==3.0.1
